restructured task view model components
