+++ wondrous-app/components/Common/TaskViewModal/constants.ts
@@ -1,0 +1,15 @@
+export const tabs = {
+  submissions: 'Submissions',
+  subTasks: 'Subtasks',
+  discussion: 'Discussion',
+  tasks: 'Tasks',
+  applications: 'Applications',
+};
+
+export const tabsPerType = {
+  proposalTabs: [tabs.discussion],
+  milestoneTabs: [tabs.tasks, tabs.discussion],
+  subtaskTabs: [tabs.submissions, tabs.discussion],
+  bountyTabs: [tabs.submissions, tabs.discussion],
+  taskTabs: [tabs.applications, tabs.submissions, tabs.subTasks, tabs.discussion],
+};
